# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/roles:
    get:
      summary: 'Get list role return json response'
      operationId: getListRoleReturnJsonResponse
      description: ''
      parameters:
        -
          in: query
          name: page
          description: 'Page number of paginated data.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'Page number of paginated data.'
            example: 1
        -
          in: query
          name: limit
          description: 'Number of items per page.'
          example: 10
          required: false
          schema:
            type: integer
            description: 'Number of items per page.'
            example: 10
        -
          in: query
          name: sort
          description: 'Sort data by created_at field.'
          example: desc
          required: false
          schema:
            type: string
            description: 'Sort data by created_at field.'
            example: desc
        -
          in: query
          name: q
          description: 'Search data by name field.'
          example: test
          required: false
          schema:
            type: string
            description: 'Search data by name field.'
            example: test
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    current_page: 1
                    data:
                      -
                        id: 3
                        name: 'Giảng viên'
                        description: null
                        created_by:
                          id: 1
                          full_name: Admin
                        updated_by:
                          id: 1
                          full_name: Admin
                        created_at: '2023-09-28T10:17:49.000000Z'
                        updated_at: '2023-09-28T10:17:49.000000Z'
                        permissions:
                          -
                            id: 1
                            name: 'Xem danh sách vai trò'
                            code: role-index
                            pivot:
                              role_id: 3
                              permission_id: 1
                          -
                            id: 2
                            name: 'Tạo mới vai trò'
                            code: role-create
                            pivot:
                              role_id: 3
                              permission_id: 2
                          -
                            id: 3
                            name: 'Xem chi tiết vai trò'
                            code: role-read
                            pivot:
                              role_id: 3
                              permission_id: 3
                      -
                        id: 2
                        name: 'Giảng viên'
                        description: null
                        created_by:
                          id: 1
                          full_name: Admin
                        updated_by:
                          id: 1
                          full_name: Admin
                        created_at: '2023-09-28T10:09:59.000000Z'
                        updated_at: '2023-09-28T10:09:59.000000Z'
                        permissions: []
                      -
                        id: 1
                        name: 'Giảng viên'
                        description: null
                        created_by:
                          id: 1
                          full_name: Admin
                        updated_by:
                          id: 1
                          full_name: Admin
                        created_at: '2023-09-28T10:09:45.000000Z'
                        updated_at: '2023-09-28T10:09:45.000000Z'
                        permissions: []
                    first_page_url: 'http://127.0.0.1:8000/api/roles?page=1'
                    from: 1
                    last_page: 1
                    last_page_url: 'http://127.0.0.1:8000/api/roles?page=1'
                    links:
                      -
                        url: null
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: 'http://127.0.0.1:8000/api/roles?page=1'
                        label: '1'
                        active: true
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    next_page_url: null
                    path: 'http://127.0.0.1:8000/api/roles'
                    per_page: 10
                    prev_page_url: null
                    to: 3
                    total: 3
                  message: success
                  status_code: 200
                properties:
                  data:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      data:
                        type: array
                        example:
                          -
                            id: 3
                            name: 'Giảng viên'
                            description: null
                            created_by:
                              id: 1
                              full_name: Admin
                            updated_by:
                              id: 1
                              full_name: Admin
                            created_at: '2023-09-28T10:17:49.000000Z'
                            updated_at: '2023-09-28T10:17:49.000000Z'
                            permissions:
                              -
                                id: 1
                                name: 'Xem danh sách vai trò'
                                code: role-index
                                pivot:
                                  role_id: 3
                                  permission_id: 1
                              -
                                id: 2
                                name: 'Tạo mới vai trò'
                                code: role-create
                                pivot:
                                  role_id: 3
                                  permission_id: 2
                              -
                                id: 3
                                name: 'Xem chi tiết vai trò'
                                code: role-read
                                pivot:
                                  role_id: 3
                                  permission_id: 3
                          -
                            id: 2
                            name: 'Giảng viên'
                            description: null
                            created_by:
                              id: 1
                              full_name: Admin
                            updated_by:
                              id: 1
                              full_name: Admin
                            created_at: '2023-09-28T10:09:59.000000Z'
                            updated_at: '2023-09-28T10:09:59.000000Z'
                            permissions: []
                          -
                            id: 1
                            name: 'Giảng viên'
                            description: null
                            created_by:
                              id: 1
                              full_name: Admin
                            updated_by:
                              id: 1
                              full_name: Admin
                            created_at: '2023-09-28T10:09:45.000000Z'
                            updated_at: '2023-09-28T10:09:45.000000Z'
                            permissions: []
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 3
                            name:
                              type: string
                              example: 'Giảng viên'
                            description:
                              type: string
                              example: null
                            created_by:
                              type: object
                              properties:
                                id:
                                  type: integer
                                  example: 1
                                full_name:
                                  type: string
                                  example: Admin
                            updated_by:
                              type: object
                              properties:
                                id:
                                  type: integer
                                  example: 1
                                full_name:
                                  type: string
                                  example: Admin
                            created_at:
                              type: string
                              example: '2023-09-28T10:17:49.000000Z'
                            updated_at:
                              type: string
                              example: '2023-09-28T10:17:49.000000Z'
                            permissions:
                              type: array
                              example:
                                -
                                  id: 1
                                  name: 'Xem danh sách vai trò'
                                  code: role-index
                                  pivot:
                                    role_id: 3
                                    permission_id: 1
                                -
                                  id: 2
                                  name: 'Tạo mới vai trò'
                                  code: role-create
                                  pivot:
                                    role_id: 3
                                    permission_id: 2
                                -
                                  id: 3
                                  name: 'Xem chi tiết vai trò'
                                  code: role-read
                                  pivot:
                                    role_id: 3
                                    permission_id: 3
                              items:
                                type: object
                                properties:
                                  id:
                                    type: integer
                                    example: 1
                                  name:
                                    type: string
                                    example: 'Xem danh sách vai trò'
                                  code:
                                    type: string
                                    example: role-index
                                  pivot:
                                    type: object
                                    properties:
                                      role_id: { type: integer, example: 3 }
                                      permission_id: { type: integer, example: 1 }
                      first_page_url:
                        type: string
                        example: 'http://127.0.0.1:8000/api/roles?page=1'
                      from:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      last_page_url:
                        type: string
                        example: 'http://127.0.0.1:8000/api/roles?page=1'
                      links:
                        type: array
                        example:
                          -
                            url: null
                            label: '&laquo; Previous'
                            active: false
                          -
                            url: 'http://127.0.0.1:8000/api/roles?page=1'
                            label: '1'
                            active: true
                          -
                            url: null
                            label: 'Next &raquo;'
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: null
                            label:
                              type: string
                              example: '&laquo; Previous'
                            active:
                              type: boolean
                              example: false
                      next_page_url:
                        type: string
                        example: null
                      path:
                        type: string
                        example: 'http://127.0.0.1:8000/api/roles'
                      per_page:
                        type: integer
                        example: 10
                      prev_page_url:
                        type: string
                        example: null
                      to:
                        type: integer
                        example: 3
                      total:
                        type: integer
                        example: 3
                  message:
                    type: string
                    example: success
                  status_code:
                    type: integer
                    example: 200
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: unauthorized
                  status_code: 401
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: unauthorized
                  status_code:
                    type: integer
                    example: 401
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: internal_server_error
                  status_code: 500
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: internal_server_error
                  status_code:
                    type: integer
                    example: 500
      tags:
        - 'Role management'
    post:
      summary: 'Create role'
      operationId: createRole
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 3
                    name: 'Giảng viên'
                    description: null
                    created_by:
                      id: 1
                      full_name: Admin
                    updated_by:
                      id: 1
                      full_name: Admin
                    created_at: '2023-09-28T10:17:49.000000Z'
                    updated_at: '2023-09-28T10:17:49.000000Z'
                    permissions:
                      -
                        id: 1
                        name: 'Xem danh sách vai trò'
                        code: role-index
                        pivot:
                          role_id: 3
                          permission_id: 1
                      -
                        id: 2
                        name: 'Tạo mới vai trò'
                        code: role-create
                        pivot:
                          role_id: 3
                          permission_id: 2
                      -
                        id: 3
                        name: 'Xem chi tiết vai trò'
                        code: role-read
                        pivot:
                          role_id: 3
                          permission_id: 3
                  message: success
                  status_code: 200
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 3
                      name:
                        type: string
                        example: 'Giảng viên'
                      description:
                        type: string
                        example: null
                      created_by:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          full_name:
                            type: string
                            example: Admin
                      updated_by:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          full_name:
                            type: string
                            example: Admin
                      created_at:
                        type: string
                        example: '2023-09-28T10:17:49.000000Z'
                      updated_at:
                        type: string
                        example: '2023-09-28T10:17:49.000000Z'
                      permissions:
                        type: array
                        example:
                          -
                            id: 1
                            name: 'Xem danh sách vai trò'
                            code: role-index
                            pivot:
                              role_id: 3
                              permission_id: 1
                          -
                            id: 2
                            name: 'Tạo mới vai trò'
                            code: role-create
                            pivot:
                              role_id: 3
                              permission_id: 2
                          -
                            id: 3
                            name: 'Xem chi tiết vai trò'
                            code: role-read
                            pivot:
                              role_id: 3
                              permission_id: 3
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            name:
                              type: string
                              example: 'Xem danh sách vai trò'
                            code:
                              type: string
                              example: role-index
                            pivot:
                              type: object
                              properties:
                                role_id:
                                  type: integer
                                  example: 3
                                permission_id:
                                  type: integer
                                  example: 1
                  message:
                    type: string
                    example: success
                  status_code:
                    type: integer
                    example: 200
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: unauthorized
                  status_code: 401
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: unauthorized
                  status_code:
                    type: integer
                    example: 401
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: internal_server_error
                  status_code: 500
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: internal_server_error
                  status_code:
                    type: integer
                    example: 500
      tags:
        - 'Role management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: q
                permissions:
                  type: array
                  description: ''
                  example:
                    - 18
                  items:
                    type: integer
              required:
                - name
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []